Improve mostly Namespace
Small changes in Entity and SharedMap.

Before, Namespace has a pointer to the actual namespace that needed to be
modified, wasting space for unncessary data. Now, all the data is put in a
NamespaceBody, which is accessed by a shared pointer, like it is for
Entity. So any copied of a Namespace will shared the same body.

The body is hidden and can only be accessed through Namespace, like it is
for Entity and its body.