Make DocumentHead and ScopedHead work properly
The entity includer didn't work (hadn't actually tested it). Was a quick
fix, it just needed to have access to Parser's entities.

Import, Scoped Document and Document Head all use the same function to get
the stuff needed.

Import only reads the contents of a file if it hasn't read them before,
otherwise it is not read.