Refactoring of the Deserializer
There's a lot to be done! The Deserializer is really not up to date with
changes in the Parser. The basic structures are mostly done. Now to finish
working on functions. The main changes that were done in the Parser relate
to fheads, not their bodies. But the code needs to be more robust
everywhere, so a whole refactoring is in order.