Namespace no access within
Having to look within a Namespace to see if there might be an Entity with
the same name makes the code much more complicated and causes unforeseen
errors. Furthermore, it makes parsing harder since it cannot be known when
the Namespace's name is parsed that it's abstract or corresponds to
another value, abstract or concrete, within the Namespace, since this is
dependent on context -- it is considered a Namespace if preceded by a
Using namespace operator, else a value is fetched within. Also, I don't
believe the potential use warrants the increase in complexity. Therefore,
the functionality has been removed.

Self reference for fhead declaration will be maintained, and perhaps to
access a value of the same name as a Namespace within it, the operator '&'
could be allowed to be appended to a Namespace's name, so the parser
easilyk known to look within.