Namespace dedicated pointer to itself
Instead of having the Namespace's pointer to itself be at the end of a
vector container any prior namespace, it'll have its own pointer.
Therefore, if it does not have any namespace prior to it, the nspaces
vector will be empty.

Having both prior namespaces and the current namespace in the same
container was a slight design error. Now they're clearly separated into
two distinct variables.

Perhaps variable and function names should be changed to make their
purpose more obvious....