Refactoring of Document Head deserialization
Progress has been made. The functions for abstract and concrete entities
have been set up, but all the deserialization they entail must still be
done.

Deserialization of Scoped Document and Import is done.

As for Using namespace, there are still problems that need to be overcame.
Notably, for the Parser, it ought to be sure that copied Entities of
different types also copy their namespaces if need be. I don't think
that's being done, so an Entity of type Webss could have many namespaces,
but its copy of another type wouldn't have any. I should rethink copying
of entities, or perhaps namespaces should be put as pointers so their type
doesn't matter...