Entities within structures now type Webss
A lot of files are affected, but in short most of changes are because
entities within structures will now be of type Webss rather than their own
type. For instance, an Entity within a BasicFunctionHead will now be
BasicEntity<Webss> instead of BasicEntity<BasicFunctionHead<Webss>>.

The rationale behind that was that having the Multi Entity Manager was
impractical, also each structure having its own entity type was also
impractical as noticed after coding of the deserialization of namespaces
was started. Copying of entities would require to copy not only the
entity, but also their namespaces. Also, it meant many types of Entities
were needed.

The reason it was this way is because, among other less important things,
that to use the template for fheads, I needed the type of the fhead. I
couldn't do getFunctionHeadStandard or getFunctionHeadBinary, etc.,
depending on the type of the fhead.

Well, that's nearly exactly what I figured out! Now, to get the fhead out
of the Webss entity, fhead calls a templated function in Webss. This
templated function converts *this to the type specified within the
template. Conversion of this calls an overloaded conversion of the type,
which returns the requested type.

This warranted the creation of Enum, to distinguish it from Namespace.
Enum is now a wrapper of Namespace and only allows insertion of names. It
takes care itself of making numbered Entities.