Using Namespace and Entity names
Now that in the parser all entities are of type Webss, it's much easier to
deserialize. No need for templates. No need to reinterpret cast Namespace
pointers. Also, no unforeseen problems.

Now there's a function "putNamespaceName" that puts the namespace's name
preceded by previous namespaces' names if necessary. To put an entity's
name, this function is called, plus adding a char scope, if necessary, and
the entity's name is put. For Using Namespace, it's just putting the char
using namespace and then calling the function.