Refactoring of Function Head deserialization
In WebssonStructures, the type Param was added to be able to get the param
type of each fhead.

Deserialization of fheads binary, standard and text was put into
templates, so a lot of duplicate code was removed. The functionality
hasn't changed though.

All that was really done is improve the structure of the code and reduce
code duplication. A lot of functionality still has to be completed and
verified in those functions.