Additional usage of templates
There's cool specialization of templates with putContainerStart and
putContainerEnd. No need for a switch statement, not even an assert, since
a function without a proper container is deleted!

A class ContainerIncluder has been made, which is like the Namespace
includer, so its constructor puts a start char and its destructor puts an
end char.

Changed ConType::Enum to auto in a lot of cases.