Deserialization of Binary data
The deserialization of binary data has been checked and refactored. There
weren't many errors, which isn't surprising since the representation of
binary data hadn't changed. Most changes were small changes to make the
code more robust, so asserts were added.

The constants to make it explicit that when there is a default value the
char is 1 and when there is none it is 0. Perhaps I'll wonder if for there
to be a default value, the char ought to be >= 1, but I think 1 or 0 is
fine.

Functions strictly related to binary deserialization were removed from the
scope of Deserializer and put exclusively in the deserializerBinary file.