Template Text now functions like Template Standard
Basically, a lot of things have been changed to make it so Template Text
is now exactly like a Template Standard, except when it comes to parsing
and deserializing. An interesting observation is that there is no need for
there to be a boolean for text stored in any structure. Whether it is text
or not is stored in Enums, so not additional space is necessary. Something
similar could easily be made for Tuples and Lists. Booleans are only
necessary in function calls, but these are dummies that are only there to
distinct the function from a standard call to a text call.

Also, Fhead in var names has been changed to Thead.