Deserialization of Templates
The deserialization of Templates has now been done with templates passing
the appropriate functions for the deserialization of tuples.

All the functions in deserializerTemplate.cpp have been moved to
deserializer.cpp, so the file has been deleted.

The functions in deserializer are all independent of the Deserializer
class. The function "putFuncBodyBinary" is called as an extern from
"putFuncBinaryTuple".