ParserBinary bugfix and WebssBinarySize check in progress
The bug has not been fixed, but improvements to the code have been made.
Among them, is that there should be no runtime error in structures. The
parser should do all the proper checks and throw runtime erros when
necessary. The bug probably occurs during the parsing of the Binary
Template head, since that's the only part that's really changed from a
while ago when the way Binary params were parsed was different.

In the Webss class, an assert has been added when getting a
WebssBinarySize.