Improvements to Binary parsing
At least two bugs were fixed. One was known and that had to do with
binary float. Now, instead of parsing the stuff into an 8 byte integer
and then casting it to float, it puts it directly in a float. Same was
done with double for consistency. Other bug was with reading a binary
size (for sizes not known in advance). It had an infinite loop. Also,
a check to make sure the read size is not greater than the maximum size
was added.

The WebssBinarySize is also properly checked in the Parser and in the
ParamBinary structure. It doesn't need to be checked anywhere else.