Got rid of global macros
Macro ALLOW_IMPORT has been changed to DISABLE_IMPORT, so it's not defined
by default. It and REVERSE_ENDIANNES have been move to the .cpp file where
they're used.

Other than that, no macro goes beyond the file scope. The file errors.h
was being mostly used by the Parser (only one of the errors was in the
Deserializer) and so has been put within its library. The files that
include it are all .cpp, so the macros aren't global.