Progress to check overflow
Overflow for integers is checked; now to check it for doubles. There is a
max number of significant digits that a double can keep; I should make it
so no more than that number of digits are checked for a double. The max
number of digits is different depending on the magnitude.