Got rid of dummy parameters
Functions with a dummy paramter are very bad! Constructors were changed so
that a static function return the type makes the job of calling the proper
constructor with a dummy parameter, which is hidden. They kind of play the
role of constructors with the same parameter, but different name and a
different result.