More efficient digit parsing
When parsing numbers, it's better to first check if the next number is a
digit rather than checking if it's whitespace. It takes more lines of
code, but the intention is clear: it is assumed in most cases a number's
digit will be unseparated by line junk, so it's just better to check if
the next character is a digit. Even for the last character after a number,
it will probably be as efficient, since in most cases the last character
will be neither a digit nor a line junk char.