Change to char escapes
Char escapes a, b, f and v have been removed.

The function for parsing string has been improved so that there is less
indirection to other functions. The same function that checks if the char
escape is valid adds the escaped char if it is, instead of using other
functions. Only the escape of Unicode chars (x, u, U and X) call another
function.