Proper deserialization of Namespace and Enum
The bug was caused because it wasn't checked if the namespace or enum had
been included; it only checked for namespaces before it. Possibly an error
caused when I decided to not keep the current namespace at the head of its
list of previous namespaces and failed to address the change everywhere.

The deserialization of scoped names has been improved as well. Before, it
looked for the first namespace in the list that hadn't been included and
put all the others. Now, the list is checked in reverse. So for instance,
for a namespace with many non-included namespaces, the previous method
would cause problems. That's one bug found and fixed.

Resolves: #29