Proper decimals parsing
Leading zeros are now skipped through, rather than being counted as
significant digits. There is a max of 23 zeros before the first
significant digit, else the decimals are considered to be equivalent to 0,
and the rest is ignored. This is similar to the max number of significant
digits. The difference is that the limit of 23 zeros is the same
regardless of the base of the number.

Resolves: #30