Bugfixes
First bug was very easy to fix: check for line escape was giving the
reverse value that it should've been giving.

The second bug was harder to fix. It was about template head text. First
of all, in one of the calls, it didn't even call the thead text
constructor, but instead the thead standard one.

Thead text is a subtype of thead standard, like text list and tuple. But
everything had been coded as if they were distinct types. Now, everything
has been fixed so that the relationship between thead text and thead
standard is coded exactly like text list and text tuple.