Comments part of line-string and bugfix
Now, for a comment to be considered as such in a line string, or in a
multiline string, it must be preceded by line junk. A line-string with the
content "http://www.[...]" will parse as such. However, something like
"http: //www.[...]" will parse as "http:" since the slashes are now
considered to be a comment.

To achieve this, since the code was well made and there was a function
that returned the next non-junk char, skipping through comments, all that
was needed to do is remove the check for comment chars in the string
parsers.

There was a typical bug in the function that parses multiline-string. The
iterator was increased when calling the function, and the function
incorrectly increased it again.

Resolves: #35