SmartIterator more constructors and copy
Copy and assignment constructors were added. Moving an iterator keeps the
pointer at the same location, whereas any copying reputs the pointer at
the start of the buffer. The same happens when a stringstream object is
passed in a constructor. However, for istream and string, it always sets
the pointer at the start of the buffer.

Resolves: #39