Usage of Tag Separator
When the separator, the getTag function returned a Tag Unknown in case of
a separator, then the char had to be checked to make sure it was a
separator. With a static separator, the separator can be put in the switch
statement of the function to get a Tag Separator.