Containers inc iterator within parse function
Before, the interator would be increased before calling most functions.
Now, when the function is called is a well-bound container, which always
has start and end char, the fonction to parse it increases the iterator.
It is then assumed that when the function is called, the iterator points
to the start char.