Abtract entity name parsing improvement
Before, the parser checked if the name was an existing entity and threw
and error if it was the case. Now, the functions that call the function to
parse an abstract entity check if the container they parse already
contain the entity or not. This allows namespaces and scoped templates to
contain already entities with names of globally defined entities within
their own scope.