Parse option
For this version (1.0.0), only one option is allowed: version. This is set
by using -v or --version followed by version name in a line-string. The
option char remains '#'.

Setting the version to 1.0.0 does nothing. The purpose is only, it there's
another version in the future, to allow a parse to know if they can or not
parse a document. The version of a document without the version option is
assumed to be 1.0.0.