Addition of binary iterator
The binary iterator will take care of manipulation bits and bytes. It is
set to increase the iterator before fetching bytes. For bits, it checks
first if the iterator must be increased. Therefore, how it parses stuff is
different than the Parser, that requires the iterator to point on the
current char, not previous one. Also, after it's done its job, the
iterator must be increased, since it'll be pointing to a previous char.