Byte block in binary iterator
The byte block allow high space efficiency when it comes to read bits.
Before, if a bit was read, then a byte, 2 whole bytes were taken, which
meant, for exemple, that 8 iterations of that would need 16 bytes. With
the byte block, the bits are read from it as long as there are bits
available, and only after that is another byte block fetched. Therefore,
the above would take 9 bytes instead of 16 bytes: 1 byte for the 8 bits,
and 8 bytes for the bytes. This makes it a bit harder for the
deserializer, but I think it is worth it.