Parser binary entity bits and bug fixes
For the size of a binary element in bits, if the '$' char precedes the
size, then it is in bits. This now also works if there's an entity after
the char. Since there can't be explicit names in a binary head, this is a
good use of the '$' char.

The previously commit was done too precipitately and had left multiples
sections of code unchecked. The code has now been thoroughly reviewed.