Refactoring
ParamDocumentIncluder was put in its own file. It needs the parser, but
doesn't need to be in the same file as the parser.

The template requirement of Import Manager has been removed. The parser
doesn't need the Import Manager in its header file. Therefore, the Import
Manager's header file can include the parser's header file, removing the
need of a template.

A static function from the curl header file was put in the curl
translation unit file.