Can't expand non-tuple within tuple
Before, lists and dictionaries could be expanded. This causes a problem in
the case of dictionaries since their keys don't have to be ordered. A
solution to this would be to make it so keys in a dictionary are ordered,
therefore a dictionary would be like a tuple with mandatory keys. Lists
don't cause a problem, but were disallowed for consistency.