Basic implementation of option alias substitution
This works for the way options are, but may need more to be more flexible
in the future. It doesn't substitute an alias that was put in another
alias's content. I think this is the way it should be. For now, an alias
is created by using ':' right after the option, then putting the name for
the alias, immediately followed by ':' and the content.