Refactor parse option and fix bugs
The way to parse the option has been changed to be a stream of tokens that
are then read. This is much better and simpler than the previous method of
reparsing a whole string.

Various bugs relating to the parsing of options and strings were fixed.
Tests were added for that.