Bugfix and add Tag Iterator
The tag iterator allows to iterate on tags and has idempotent access to
the current tag. If the tag iterator is deemed to be invalid, or it points
to junk, it gets the next tag, else it gets the last tag it fetched. This
avoids issues with text containers, which was causing the bug.