Improve parser tag iterator
The tag iterator is well defined now, so its comprehension is very easy,
leading to less problems. Now it's clear that there's a parsing for tags,
and that there's a parsing for chars. They should not be mixed together.
For instance, after the end of a container, instead of merely increasing
the chars iterator, the tag iterator is increased.