Import as expand of document
A document is now imported as if it were in the current document as is.
Since key-values from the imported document are added to the current
document's body, a new alternate type of "import" has been added. This
allows the serializer to know which values in a document body were
imported so that it can ignore them.

Resolves: #16