Change place of switch default for assert
The default in a switch that leads to an error is used to avoid compiler
errors. In the case of functions that return a value, a throw statement
was also added after an "assert(false)" so that there'd be no compiler
errors when the code was compiled without allowing assertions. This
would unnecessarily bloat the code for something that shouldn't happen.
Instead, the default is just put above another statement that does
something. The behavior then really is undefined.