Bugfixes parse strings
When parsing a string start, the iterator is set as unsafe. However,
when there is `${...}`, then, since it parses a value that may be an
entity, it can become safe again, and therefore has to be made unsafe to
avoid errors.

This is an extremely hard bug to catch. Function `parseValueOnly` calls
`parseValueOther`, which calls `parseNameType`, where it may call
`tagit.update()`, which makes the iterator safe again.